+++ src/instructlab/model/backends/backends.py
@@ -2,10 +2,8 @@
 
 # Standard
 from typing import Optional, Tuple
-import json
 import logging
 import pathlib
-import struct
 import sys
 
 # Third Party
@@ -13,92 +11,13 @@ import click
 
 # Local
 from ...configuration import _serve as serve_config
-from ...utils import split_hostport
+from ...utils import is_model_gguf, is_model_safetensors, split_hostport
 from .common import CHAT_TEMPLATE_AUTO, LLAMA_CPP, VLLM
 from .server import BackendServer
 
 logger = logging.getLogger(__name__)
 
 SUPPORTED_BACKENDS = frozenset({LLAMA_CPP, VLLM})
-
-
-def is_model_safetensors(model_path: pathlib.Path) -> bool:
-    """Check if model_path is a valid safe tensors directory
-
-    Check if provided path to model represents directory containing a safetensors representation
-    of a model. Directory must contain a specific set of files to qualify as a safetensors model directory
-    Args:
-        model_path (Path): The path to the model directory
-    Returns:
-        bool: True if the model is a safetensors model, False otherwise.
-    """
-    try:
-        files = list(model_path.iterdir())
-    except (FileNotFoundError, NotADirectoryError, PermissionError) as e:
-        logger.debug("Failed to read directory: %s", e)
-        return False
-
-    # directory should contain either .safetensors or .bin files to be considered valid
-    if not any(file.suffix in (".safetensors", ".bin") for file in files):
-        logger.debug("'%s' has no *.safetensors or *.bin files", model_path)
-        return False
-
-    basenames = {file.name for file in files}
-    requires_files = {
-        "config.json",
-        "tokenizer.json",
-        "tokenizer_config.json",
-    }
-    diff = requires_files.difference(basenames)
-    if diff:
-        logger.debug("'%s' is missing %s", model_path, diff)
-        return False
-
-    for file in model_path.glob("*.json"):
-        try:
-            with file.open(encoding="utf-8") as f:
-                json.load(f)
-        except (PermissionError, json.JSONDecodeError) as e:
-            logger.debug("'%s' is not a valid JSON file: e", file, e)
-            return False
-
-    # TODO: add check for safetensors file header (?)
-    return True
-
-
-def is_model_gguf(model_path: pathlib.Path) -> bool:
-    """
-    Check if the file is a GGUF file.
-    Args:
-        model_path (Path): The path to the file.
-    Returns:
-        bool: True if the file is a GGUF file, False otherwise.
-    """
-    # Third Party
-    from gguf.constants import GGUF_MAGIC
-
-    try:
-        with model_path.open("rb") as f:
-            first_four_bytes = f.read(4)
-
-        # Convert the first four bytes to an integer
-        first_four_bytes_int = int(struct.unpack("<I", first_four_bytes)[0])
-
-        return first_four_bytes_int == GGUF_MAGIC
-    except struct.error as exc:
-        logger.debug(
-            f"Failed to unpack the first four bytes of {model_path}. "
-            f"The file might not be a valid GGUF file or is corrupted: {exc}"
-        )
-        return False
-    except IsADirectoryError as exc:
-        logger.debug(f"GGUF Path {model_path} is a directory, returning {exc}")
-        return False
-    except OSError as exc:
-        logger.debug(
-            f"An unexpected error occurred while processing {model_path}: {exc}"
-        )
-        return False
 
 
 def determine_backend(model_path: pathlib.Path) -> Tuple[str, str]:
