+++ src/instructlab/model/evaluate.py
@@ -18,7 +18,7 @@ from instructlab import clickext
 from instructlab.model.backends import backends
 
 # Local
-from ..utils import http_client
+from ..utils import http_client, is_model_gguf, is_model_safetensors
 
 logger = logging.getLogger(__name__)
 
@@ -118,9 +118,9 @@ def validate_model(model: str, model_arg: str = "--model"):
         model_path = pathlib.Path(model)
         valid_model = False
         if model_path.is_dir():
-            valid_model = backends.is_model_safetensors(model_path)
+            valid_model = is_model_safetensors(model_path)
         elif model_path.is_file():
-            valid_model = backends.is_model_gguf(model_path)
+            valid_model = is_model_gguf(model_path)
         if not valid_model:
             click.secho(
                 f"Evaluate '{model_arg}' needs to be passed either a safetensors directory or a GGUF file",
